bug: allow scan_consistency on query options
This solves #53, but does not completely implement all the query options
for the 3.0 SDK.

It lets you set the scan_consistency to either NotBounded or
RequestPlus. The logic for passing in a mutation token and getting the
scscan_consistency set to at_plus is left for another PR, along with
the other options.

Signed-off-by: Adam Jacob <adam@stalecoffee.org>