Move to nightly rust to access async closures
Motivation
----------
A primary design pattern that we want to use is currently challenging
due to asybc closures not being supported in stable rust. In order
to make this pattern possible we should switch to nightly and enable
the feature. async_closures should be moved to stable before we need
to release the SDK.

Changes
--------
Specify the nightly toolchain.
Enable async_closures feature.
Use async closures.