Change authenticator to be an enum
Motivation
----------
Authenticator being a trait makes it more difficult to work with
than if it was an enum. It is intended to be a public facing
interface but with private-only implementations. As we know all
of the implementations we can just use an enum.

Changes
--------
Change authenticator to be an enum.