Rework json stream implementation
Motivation
----------
We currently implement stream so that it streams all of the rows
and then returns None, a user must then cfall read_epilog to finish
reading items off the stream - e.g. metadata. This is actually
quite challenging to do as a consumer of the json stream, it'd
be better if the epilog was parsed as a part of the stream.