RSCBC-64: Make opening a bucket a background task
Motivation
----------
The call to cluster.bucket is supposed to be synchronous and not
return a result. At present it's async and has an unwrap in it.

Changes
-------
Add an agent provider which spawns a task to get an agent at the
bucket level.
Obtain agents in clients via the provider.