RSCBC-87: Remove retries from public API
Motivation
----------
The current state of the retry RFC does not really achieve
functionality that users wants. Retries are also one of the
most difficult parts of our API to design. For now we should
remove support from the public API but leave functionality in core
so that we behave the same as other SDKs for default retry behaviours.

Changes
-------
Remove retries from public API.