RSCBC-100: Attempt to parse address in kv as DNS name first for rustls
Motivation
----------
We currently always parse addresses into IP before connecting over
memcached. This means that certificate validation can fail because
the IP is not listed as a subject in the certificate, instead the
DNS server name is in the cert.

Changes
-------
Update core SeedConfig to accept a new Address type so that we can
easily track bootstrap hosts and posts separetely.
When creating the rustls ServerName first attempt to use the
host as DNSName, falling back to the socket's peer address IP if
that fails.
Address RSCBC-112 at the same time, moving from certificates in
the public API being byte vectors to being relevant certificate
types.